runtime/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-runtime"
version = "0.1.0a1"
description = "MCP Runtime Orchestrator"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "MCP Platform Team"},
]
keywords = ["mcp", "api", "runtime", "orchestrator"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.5.0,<3.0.0",
    "pydantic-settings>=2.1.0,<3.0.0",
    "fastapi>=0.104.0,<1.0.0",
    "uvicorn[standard]>=0.24.0,<1.0.0",
    "jsonschema>=4.20.0,<5.0.0",
    "PyYAML>=6.0.1,<7.0.0",
    "httpx>=0.25.0,<1.0.0",
    "websockets>=12.0,<13.0",
    "azure-keyvault-secrets>=4.7.0,<5.0.0",
    "azure-identity>=1.15.0,<2.0.0",
    "click>=8.1.0,<9.0.0",
    "python-json-logger>=3.0.0,<4.0.0",
    "cryptography>=41.0.0,<42.0.0",
    "openapi-spec-validator>=0.7.0,<1.0.0",
    "prance>=23.6.0.0,<24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0,<8.0.0",
    "pytest-asyncio>=0.21.0,<1.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "black>=23.0.0,<24.0.0",
    "isort>=5.12.0,<6.0.0",
    "mypy>=1.7.0,<2.0.0",
]

[project.urls]
Homepage = "https://github.com/alonilk2/openapi-mcp-connector"
Repository = "https://github.com/alonilk2/openapi-mcp-connector"
Issues = "https://github.com/alonilk2/openapi-mcp-connector/issues"

[project.scripts]
# CLI for working with connectors/manifests
mcp = "runtime.cli.main:cli"
# Run the FastAPI server via uvicorn
mcp-runtime = "runtime.run:main"

[tool.black]
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=runtime",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "-v"
]
asyncio_mode = "auto"

[tool.hatch.build]
include = [
    "runtime/**",
    "README.md",
]
exclude = [
    "runtime/tests/**",
    "**/__pycache__/**",
    "**/.pytest_cache/**",
]
